requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
openpyxl==3.0.10
Pillow==9.3.0
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.20.0
typing-extensions==4.4.0
yarl==1.8.1
